Expose keeaplive settings
This change allows to set some TCP socket options during the creation of
the transport.
The parameters exposed are:
- TCP_KEEPIDLE
- TCP_KEEPINTVL
- TCP_KEEPCNT

Those values can be set passing a dictionray where the key is the name of
the parameter we want to set.
The names of the keys are the ones reported above.