Expose keepalive settings
This change allows to set some TCP socket options during the creation of
the transport.
The parameters we expose with this change are:
- TCP_KEEPIDLE
- TCP_KEEPINTVL
- TCP_KEEPCNT

Those values can be set passing a dictionary where the key is tha name of
the parameter and the value is its value. The names of the keys are the
ones reported above.