Set tcp socket options
This change allows to set the TCP options for the sockets used by
transport class. The socket options are passed as a dictionary where the
key is the name of the option and the value is set as option value. At
the moment we support just options for the TCP level.