Add a 'nowait' parameter to Connection.close()
If the client is running connection.drain_events() in another thread,
there is no good way to terminate the connection when connection.close()
waits for the reply: two readers on one connection will either raise an
exception or deadlock.

Setting nowait=True in this situation will let the other thread read the
reply and cleanly close the conection.