do not reconnect on collect (#121)
* do not reconnect on collect

collect is called to cleanup when there is a connection error but
the backward compatibility code to connect when transport is called
gets triggered. this doesn't seem right as this should be handled by
ensure_connection.

* do not reconnect on collect

collect is called to cleanup when there is a connection error but
the backward compatibility code to connect when transport is called
gets triggered. this doesn't seem right as this should be handled by
ensure_connection.