win32: TCP_MAXSEG isn't available for setsockopt
According to [1,2] Windows platforms support getsockopt(TCP_MAXSSEG) but not setsockopt(TCP_MAXSEG) on IPPROTO_TCP sockets.  This results in:

socket.error: [Errno 10042] An unknown, invalid, or unsupported option or level was specified in a getsockopt or setsockopt call

when trying to connect to the server.

This is a latent issue uncovered by the refactor in:
    Set default KEEPIDLE/KEEPCNT/KEEPINTVL and TCP_USER_TIMEOUT (b14962c2dc)

[1] https://msdn.microsoft.com/en-us/library/windows/desktop/ms740476(v=vs.85).aspx
[2] https://msdn.microsoft.com/en-us/library/windows/desktop/ms738544(v=vs.85).aspx