Fix a channel issue against a connection already closed
In some possible case a connection can be closed (network issue, etc...) and
then channels reset to None, in this case if the client use
the on_inbound_method then we can facing a TypeError exception.

The `on_inbound_method` is not enough safer and then this method can
try to get channel from a none existing object.

In this case we need to raise a `RecoverableConnectionError` to allow client
to re-connect properly and re-init objects properly.