Fix pypi classifier development status
The version 5.0 is currently an alpha version
but pypi classifiers tag this version as a production/stable version,
this is wrong, this version is a pre-alpha version.

The main issue is that by using pip or requirements file to install
py-amqp it will pull and install the pre-alpha version (5.0) instead of
the stable version 2.5.2.

Normally pip don't install non stable version except if user
explicitly specify to install a non stable version.

Example with oslo.messaging:

```
$ pip install oslo.messaging
```

The previous command will pull py-amqp 5.0