Fix saving partial read buffer on windows during socket timeout. (#321)
* Fix saving partial read buffer on windows during socket timeout.

See Issue #320.

When a frame read times out on Windows, the error raised did not
trigger the transport logic to save off the initial partial read
so that a future call to `read_frame()` would reread the partially
read data.  This caused framing errors in certain cases on Windows.

Also fixes a similar error that could occur on any platform during
reads on payloads > SIGNED_INT_MAX, where reading the payload is
not an atomic read but split into two reads.  If the first succeeded
but the second failed, the first read would be thrown away causing
framing errors on subsequent calls to `read_frame`.

* Don't assume existence of BlockingIOError in tests

* Refactor comments and narrow exception around partial payload read